Added new methods to DataQueryServiceBean

+++ CB004273_ClassLibrary/src/kavith/jee/assignment/service/DataQuerySerivceBeanRemote.java
@@ -24,6 +24,6 @@ public interface DataQuerySerivceBeanRemote {
     BookingDetails geBookingDetailsById(String id);
     List<BookingDetails> getBookingsByPassenger(String passengerID);
     List<BookingDetails> getBookingsByFlight(String flightID);    
-    Map<String,Integer> getAvailableNoOfSeatsInAllFlights();
+    Map<String,Short> getAvailableNoOfSeatsInAllFlights();
     List<AircraftDetails> getAllUnallocatedAircrafts();
 }
